Moving SMS exclusive text properties to a exclusive SMS text content
