+++ spec/components/schemas/content/sms/all.ts
@@ -1,5 +1,5 @@
 import { SchemaObject } from 'openapi3-ts';
-import { ref as textRef } from '../text';
+import { ref as textRef } from './text';
 import { createComponentRef } from '../../../../../utils/ref';
 
 const all: SchemaObject = {
